LayoutTests/imported/w3c/web-platform-tests/resource-timing/resource_TAO_match_wildcard-expected.txt

Blocked access to external URL http://www.localhost:8800/resource-timing/resources/TAOResponse.py?tao=match_wildcard
CONSOLE MESSAGE: line 13: NetworkError:  A network error occurred.
Description

This test validates that for a cross origin resource, the timing allow check algorithm will pass when the Timing-Allow-Origin header value list contains a wildcard ("*").


PASS window.performance is defined 
PASS window.performance.getEntriesByType is defined 
